Rename package name to Extensions
