+++ api/src/main/kotlin/com/example/todo/application/handler/TodoExceptionHandler.kt
@@ -1,7 +1,7 @@
 package com.example.todo.application.handler
 
 import com.example.todo.application.reponse.ResponseData
-import com.example.todo.commons.log
+import com.example.todo.extentions.log
 import com.example.todo.domain.exception.NotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ExceptionHandler
