+++ api/src/test/kotlin/com/example/todo/commons/ExtensionsKtTest.kt
@@ -1,5 +1,6 @@
 package com.example.todo.commons
 
+import com.example.todo.extentions.httpStatus
 import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
